pyproject.toml
[project]
name = "pyenv-st-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.58",
    "langchain-text-splitters>=0.3.8",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.0",
    "tiktoken>=0.9.0",
]
